gtk4: Manually implement `GraphicsOffload` constructor for now
Needs to be transfer-none instead of transfer-full
